Starting to separate the C# tensor implementation into multiple files.
