+++ src/TorchSharp/Tensor/TorchTensor.cs
@@ -44,7 +44,7 @@ namespace TorchSharp.Tensor
     /// <summary>
     /// Represents a Torch tensor.
     /// </summary>
-    public sealed class TorchTensor : IDisposable
+    public sealed partial class TorchTensor : IDisposable
     {
         internal IntPtr handle;
 
@@ -1289,335 +1289,6 @@ namespace TorchSharp.Tensor
             }
         }
 
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_add(IntPtr tensor, IntPtr trg, IntPtr alpha);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="target"></param>
-        /// <returns></returns>
-        public TorchTensor add(TorchTensor target)
-        {
-            return add(target, 1);
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="target"></param>
-        /// <param name="alpha"></param>
-        /// <returns></returns>
-        public TorchTensor add(TorchTensor target, TorchScalar alpha)
-        {
-            var res = THSTensor_add(handle, target.Handle, alpha.Handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_add_scalar(IntPtr tensor, IntPtr trg, IntPtr alpha);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="scalar"></param>
-        /// <returns></returns>
-        public TorchTensor add(TorchScalar scalar)
-        {
-            return add(scalar, 1);
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="scalar"></param>
-        /// <param name="alpha"></param>
-        /// <returns></returns>
-        public TorchTensor add(TorchScalar scalar, TorchScalar alpha)
-        {
-            return new TorchTensor(THSTensor_add_scalar(handle, scalar.Handle, alpha.Handle));
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_add_(IntPtr tensor, IntPtr trg, IntPtr alpha);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="target"></param>
-        /// <returns></returns>
-        public TorchTensor add_(TorchTensor target)
-        {
-            return add_(target, 1);
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="target"></param>
-        /// <param name="alpha"></param>
-        /// <returns></returns>
-        public TorchTensor add_(TorchTensor target, TorchScalar alpha)
-        {
-            return new TorchTensor(THSTensor_add_(handle, target.Handle, alpha.Handle));
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_add_scalar_(IntPtr tensor, IntPtr trg, IntPtr alpha);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="scalar"></param>
-        /// <returns></returns>
-        public TorchTensor add_(TorchScalar scalar)
-        {
-            return add_(scalar, 1);
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="scalar"></param>
-        /// <param name="alpha"></param>
-        /// <returns></returns>
-        public TorchTensor add_(TorchScalar scalar, TorchScalar alpha)
-        {
-            var res = THSTensor_add_scalar_(handle, scalar.Handle, alpha.Handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_addbmm(IntPtr mat, IntPtr batch1, IntPtr batch2, float beta, float alpha);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="batch1"></param>
-        /// <param name="batch2"></param>
-        /// <param name="beta"></param>
-        /// <param name="alpha"></param>
-        /// <returns></returns>
-        public TorchTensor addbmm(TorchTensor batch1, TorchTensor batch2, float beta = 1, float alpha = 1)
-        {
-            var res = THSTensor_addbmm(handle, batch1.Handle, batch2.Handle, beta, alpha);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_addbmm_(IntPtr mat, IntPtr batch1, IntPtr batch2, float beta, float alpha);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="batch1"></param>
-        /// <param name="batch2"></param>
-        /// <param name="beta"></param>
-        /// <param name="alpha"></param>
-        /// <returns></returns>
-        public TorchTensor addbmm_(TorchTensor batch1, TorchTensor batch2, float beta = 1, float alpha = 1)
-        {
-            var res = THSTensor_addbmm_(handle, batch1.Handle, batch2.Handle, beta, alpha);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_addcdiv(IntPtr tensor, IntPtr tensor1, IntPtr tensor2, IntPtr value);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="tensor1"></param>
-        /// <param name="tensor2"></param>
-        /// <param name="value"></param>
-        /// <returns></returns>
-        public TorchTensor addcdiv(TorchTensor tensor1, TorchTensor tensor2, TorchScalar value)
-        {
-            var res = THSTensor_addcdiv(handle, tensor1.Handle, tensor2.Handle, value.Handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_addcdiv_(IntPtr tensor, IntPtr tensor1, IntPtr tensor2, IntPtr value);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="tensor1"></param>
-        /// <param name="tensor2"></param>
-        /// <param name="value"></param>
-        /// <returns></returns>
-        public TorchTensor addcdiv_(TorchTensor tensor1, TorchTensor tensor2, TorchScalar value)
-        {
-            var res = THSTensor_addcdiv_(handle, tensor1.Handle, tensor2.Handle, value.Handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_addcmul(IntPtr tensor, IntPtr tensor1, IntPtr tensor2, IntPtr value);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="tensor1"></param>
-        /// <param name="tensor2"></param>
-        /// <param name="value"></param>
-        /// <returns></returns>
-        public TorchTensor addcmul(TorchTensor tensor1, TorchTensor tensor2, TorchScalar value)
-        {
-            var res = THSTensor_addcmul(handle, tensor1.Handle, tensor2.Handle, value.Handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_addcmul_(IntPtr tensor, IntPtr tensor1, IntPtr tensor2, IntPtr value);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="tensor1"></param>
-        /// <param name="tensor2"></param>
-        /// <param name="value"></param>
-        /// <returns></returns>
-        public TorchTensor addcmul_(TorchTensor tensor1, TorchTensor tensor2, TorchScalar value)
-        {
-            var res = THSTensor_addcmul_(handle, tensor1.Handle, tensor2.Handle, value.Handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_addmm(IntPtr mat, IntPtr mat1, IntPtr mat2, float beta, float alpha);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="mat1"></param>
-        /// <param name="mat2"></param>
-        /// <param name="beta"></param>
-        /// <param name="alpha"></param>
-        /// <returns></returns>
-        public TorchTensor addmm(TorchTensor mat1, TorchTensor mat2, float beta, float alpha)
-        {
-            var res = THSTensor_addmm(handle, mat1.Handle, mat2.Handle, beta, alpha);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_addmm_(IntPtr mat, IntPtr mat1, IntPtr mat2, float beta, float alpha);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="mat1"></param>
-        /// <param name="mat2"></param>
-        /// <param name="beta"></param>
-        /// <param name="alpha"></param>
-        /// <returns></returns>
-        public TorchTensor addmm_(TorchTensor mat1, TorchTensor mat2, float beta, float alpha)
-        {
-            var res = THSTensor_addmm_(handle, mat1.Handle, mat2.Handle, beta, alpha);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_addmv(IntPtr mat, IntPtr mat1, IntPtr vec2, float beta, float alpha);
-
-        /// <summary>
-        /// Performs a matrix-vector product of the matrix mat and the vector vec. The vector input is added to the final result.
-        /// </summary>
-        /// <param name="vec1"></param>
-        /// <param name="vec2"></param>
-        /// <param name="beta"></param>
-        /// <param name="alpha"></param>
-        /// <returns></returns>
-        public TorchTensor addmv(TorchTensor vec1, TorchTensor vec2, float beta = 1.0f, float alpha = 1.0f)
-        {
-            var res = THSTensor_addmv(handle, vec1.Handle, vec2.Handle, beta, alpha);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_addmv_(IntPtr mat, IntPtr mat1, IntPtr vec2, float beta, float alpha);
-
-        /// <summary>
-        /// Performs a matrix-vector product of the matrix mat and the vector vec. The vector input is added to the final result.
-        /// </summary>
-        /// <param name="vec1"></param>
-        /// <param name="vec2"></param>
-        /// <param name="beta"></param>
-        /// <param name="alpha"></param>
-        /// <returns></returns>
-        public TorchTensor addmv_(TorchTensor vec1, TorchTensor vec2, float beta = 1.0f, float alpha = 1.0f)
-        {
-            var res = THSTensor_addmv_(handle, vec1.Handle, vec2.Handle, beta, alpha);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_addr(IntPtr mat, IntPtr mat1, IntPtr vec2, float beta, float alpha);
-
-        /// <summary>
-        /// Performs the outer-product of vectors vec1 and vec2 and adds it to the input tensor.
-        /// </summary>
-        /// <param name="vec1"></param>
-        /// <param name="vec2"></param>
-        /// <param name="beta"></param>
-        /// <param name="alpha"></param>
-        /// <returns></returns>
-        public TorchTensor addr(TorchTensor vec1, TorchTensor vec2, float beta = 1.0f, float alpha = 1.0f)
-        {
-            var res = THSTensor_addr(handle, vec1.Handle, vec2.Handle, beta, alpha);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_addr_(IntPtr mat, IntPtr mat1, IntPtr vec2, float beta, float alpha);
-
-        /// <summary>
-        /// Performs the outer-product of vectors vec1 and vec2 and adds it to the input tensor.
-        ///
-        /// In-place version of 'addr'
-        /// </summary>
-        /// <param name="vec1"></param>
-        /// <param name="vec2"></param>
-        /// <param name="beta"></param>
-        /// <param name="alpha"></param>
-        /// <returns></returns>
-        public TorchTensor addr_(TorchTensor vec1, TorchTensor vec2, float beta = 1.0f, float alpha = 1.0f)
-        {
-            var res = THSTensor_addr_(handle, vec1.Handle, vec2.Handle, beta, alpha);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_all(IntPtr tensor);
@@ -1793,291 +1464,6 @@ namespace TorchSharp.Tensor
         }
 
         [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_cos(IntPtr tensor);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <returns></returns>
-        public TorchTensor cos()
-        {
-            var res = THSTensor_cos(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_cos_(IntPtr tensor);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <returns></returns>
-        public TorchTensor cos_()
-        {
-            var res = THSTensor_cos_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_sin(IntPtr tensor);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <returns></returns>
-        public TorchTensor sin()
-        {
-            var res = THSTensor_sin(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_sin_(IntPtr tensor);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <returns></returns>
-        public TorchTensor sin_()
-        {
-            var res = THSTensor_sin_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_tan(IntPtr tensor);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <returns></returns>
-        public TorchTensor tan()
-        {
-            var res = THSTensor_tan(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_tan_(IntPtr tensor);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <returns></returns>
-        public TorchTensor tan_()
-        {
-            var res = THSTensor_tan_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_angle(IntPtr tensor);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <returns></returns>
-        public TorchTensor angle()
-        {
-            return new TorchTensor(THSTensor_angle(handle));
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_asin(IntPtr tensor);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <returns></returns>
-        public TorchTensor asin()
-        {
-            return new TorchTensor(THSTensor_asin(handle));
-        }
-
-        public TorchTensor arcsin() => asin();
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_asin_(IntPtr tensor);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <returns></returns>
-        public TorchTensor asin_()
-        {
-            var res = THSTensor_asin_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        public TorchTensor arcsin_() => asin_();
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_acos(IntPtr tensor);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <returns></returns>
-        public TorchTensor acos()
-        {
-            var res = THSTensor_acos(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        public TorchTensor arccos() => acos();
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_acos_(IntPtr tensor);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <returns></returns>
-        public TorchTensor acos_()
-        {
-            var res = THSTensor_acos_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        public TorchTensor arccos_() => acos_();
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_atan(IntPtr tensor);
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <returns></returns>
-        public TorchTensor atan()
-        {
-            var res = THSTensor_atan(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        public TorchTensor arctan() => atan();
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_atan_(IntPtr tensor);
-
-        public TorchTensor atan_()
-        {
-            var res = THSTensor_atan_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        public TorchTensor arctan_() => atan_();
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_atan2(IntPtr tensor, IntPtr other);
-
-        public TorchTensor atan2(TorchTensor other)
-        {
-            var res = THSTensor_atan2(handle, other.Handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_atan2_(IntPtr tensor, IntPtr other);
-
-        public TorchTensor atan2_(TorchTensor other)
-        {
-            var res = THSTensor_atan2_(handle, other.Handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_sinc(IntPtr tensor);
-
-        public TorchTensor sinc()
-        {
-            var res = THSTensor_sinc(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_sinc_(IntPtr tensor);
-
-        public TorchTensor sinc_()
-        {
-            var res = THSTensor_sinc_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_sinh(IntPtr tensor);
-
-        public TorchTensor sinh()
-        {
-            var res = THSTensor_sinh(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_sinh_(IntPtr tensor);
-
-        public TorchTensor sinh_()
-        {
-            var res = THSTensor_sinh_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_cosh(IntPtr tensor);
-
-        public TorchTensor cosh()
-        {
-            var res = THSTensor_cosh(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_cosh_(IntPtr tensor);
-
-        public TorchTensor cosh_()
-        {
-            var res = THSTensor_cosh_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_count_nonzero(IntPtr tensor, IntPtr dim, int dim_len);
 
         public TorchTensor count_nonzero(long[]? dims = null)
@@ -2104,128 +1490,6 @@ namespace TorchSharp.Tensor
                     return new TorchTensor(res);
                 }
             }
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_tanh(IntPtr tensor);
-
-        public TorchTensor tanh()
-        {
-            var res = THSTensor_tanh(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_tanh_(IntPtr tensor);
-
-        public TorchTensor tanh_()
-        {
-            var res = THSTensor_tanh_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_arcsinh(IntPtr tensor);
-
-        public TorchTensor arcsinh()
-        {
-            var res = THSTensor_arcsinh(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_arcsinh_(IntPtr tensor);
-
-        public TorchTensor arcsinh_()
-        {
-            var res = THSTensor_arcsinh_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_arccosh(IntPtr tensor);
-
-        public TorchTensor arccosh()
-        {
-            var res = THSTensor_arccosh(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_arccosh_(IntPtr tensor);
-
-        public TorchTensor arccosh_()
-        {
-            var res = THSTensor_arccosh_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_arctanh(IntPtr tensor);
-
-        public TorchTensor arctanh()
-        {
-            var res = THSTensor_arctanh(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_arctanh_(IntPtr tensor);
-
-        public TorchTensor arctanh_()
-        {
-            var res = THSTensor_arctanh_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        public TorchTensor asinh() => arcsinh();
-
-        public TorchTensor asinh_() => arctanh_();
-
-        public TorchTensor acosh() => arccosh();
-
-        public TorchTensor acosh_() => arccosh_();
-
-        public TorchTensor atanh() => arctanh();
-
-        public TorchTensor atanh_() => arctanh_();
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_floor(IntPtr tensor);
-
-        public TorchTensor floor()
-        {
-            var res = THSTensor_floor(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_floor_(IntPtr tensor);
-
-        public TorchTensor floor_()
-        {
-            var res = THSTensor_floor_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
         }
 
         [DllImport("LibTorchSharp")]
@@ -2317,61 +1581,6 @@ namespace TorchSharp.Tensor
         }
 
         [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_ceil(IntPtr tensor);
-
-        public TorchTensor ceil()
-        {
-            var res = THSTensor_ceil(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_ceil_(IntPtr tensor);
-
-        public TorchTensor ceil_()
-        {
-            var res = THSTensor_ceil_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_sign(IntPtr tensor);
-
-        public TorchTensor sign()
-        {
-            var res = THSTensor_sign(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_sign_(IntPtr tensor);
-
-        public TorchTensor sign_()
-        {
-            var res = THSTensor_sign_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_signbit(IntPtr tensor);
-
-        public TorchTensor signbit()
-        {
-            var res = THSTensor_signbit(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_softplus(IntPtr tensor);
 
         public TorchTensor softplus()
@@ -2410,28 +1619,6 @@ namespace TorchSharp.Tensor
         public TorchTensor relu_()
         {
             var res = THSTensor_relu_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_reciprocal(IntPtr tensor);
-
-        public TorchTensor reciprocal()
-        {
-            var res = THSTensor_reciprocal(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_reciprocal_(IntPtr tensor);
-
-        public TorchTensor reciprocal_()
-        {
-            var res = THSTensor_reciprocal_(handle);
             if (res == IntPtr.Zero)
                 Torch.CheckForErrors();
             return new TorchTensor(res);
@@ -2592,17 +1779,6 @@ namespace TorchSharp.Tensor
         }
 
         [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_hypot(IntPtr tensor, IntPtr other);
-
-        public TorchTensor hypot(TorchTensor other)
-        {
-            var res = THSTensor_hypot(handle, other.handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_igamma(IntPtr tensor, IntPtr other);
 
         public TorchTensor igamma(TorchTensor other)
@@ -2724,28 +1900,6 @@ namespace TorchSharp.Tensor
         }
 
         [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_round(IntPtr tensor);
-
-        public TorchTensor round()
-        {
-            var res = THSTensor_round(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_round_(IntPtr tensor);
-
-        public TorchTensor round_()
-        {
-            var res = THSTensor_round_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_selu(IntPtr tensor);
 
         public TorchTensor selu()
@@ -2791,103 +1945,11 @@ namespace TorchSharp.Tensor
         }
 
         [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_abs(IntPtr tensor);
-
-        public TorchTensor abs()
-        {
-            var res = THSTensor_abs(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        public TorchTensor absolute() => abs();
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_abs_(IntPtr tensor);
-
-        public TorchTensor abs_()
-        {
-            var res = THSTensor_abs_(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        public TorchTensor absolute_() => abs_();
-
-        [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_log_sigmoid(IntPtr tensor);
 
         public TorchTensor log_sigmoid()
         {
             var res = THSTensor_log_sigmoid(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_logaddexp(IntPtr tensor, IntPtr other);
-
-        public TorchTensor logaddexp(TorchTensor other)
-        {
-            var res = THSTensor_logaddexp(handle, other.handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_logaddexp2(IntPtr tensor, IntPtr other);
-
-        public TorchTensor logaddexp2(TorchTensor other)
-        {
-            var res = THSTensor_logaddexp2(handle, other.handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_logcumsumexp(IntPtr tensor, long dim);
-
-        public TorchTensor logcumsumexp(long dim)
-        {
-            var res = THSTensor_logcumsumexp(handle, dim);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_logsumexp(IntPtr tensor, long dim, bool keepdim);
-
-        public TorchTensor logsumexp(long dim, Boolean keepdim = false)
-        {
-            var res = THSTensor_logsumexp(handle, dim, keepdim);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_log10(IntPtr tensor);
-
-        public TorchTensor log10()
-        {
-            var res = THSTensor_log10(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_log10_(IntPtr tensor);
-
-        public TorchTensor log10_()
-        {
-            var res = THSTensor_log10_(handle);
             if (res == IntPtr.Zero)
                 Torch.CheckForErrors();
             return new TorchTensor(res);
@@ -2915,130 +1977,6 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_log1p(IntPtr tensor);
-
-        public TorchTensor log1p()
-        {
-            var res = THSTensor_log1p(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_log1p_(IntPtr tensor);
-
-        public TorchTensor log1p_()
-        {
-            var res = THSTensor_log1p_(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_log2(IntPtr tensor);
-
-        public TorchTensor log2()
-        {
-            var res = THSTensor_log2(handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_log2_(IntPtr tensor);
-
-        public TorchTensor log2_()
-        {
-            var res = THSTensor_log2_(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_logit(IntPtr tensor, IntPtr eps);
-
-        public TorchTensor logit(double? eps = null)
-        {
-            var epsArr = eps.HasValue ? new double[] { eps.Value } : null;
-
-            unsafe {
-                fixed (double* pEps = epsArr) {
-                    var res = THSTensor_logit(handle, (IntPtr)pEps);
-                    if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-                    return new TorchTensor(res);
-                }
-            }
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_sqrt(IntPtr tensor);
-
-        public TorchTensor sqrt()
-        {
-            var res = THSTensor_sqrt(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_sqrt_(IntPtr tensor);
-
-        public TorchTensor sqrt_()
-        {
-            var res = THSTensor_sqrt_(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_rsqrt(IntPtr tensor);
-
-        public TorchTensor rsqrt()
-        {
-            var res = THSTensor_rsqrt(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_rsqrt_(IntPtr tensor);
-
-        public TorchTensor rsqrt_()
-        {
-            var res = THSTensor_rsqrt_(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        public static TorchTensor operator -(TorchTensor tensor)
-        {
-            return tensor.neg();
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_neg(IntPtr tensor);
-
-        public TorchTensor neg()
-        {
-            var res = THSTensor_neg(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        public TorchTensor negative() => neg();
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_neg_(IntPtr tensor);
-
-        public TorchTensor neg_()
-        {
-            var res = THSTensor_neg_(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_baddbmm(IntPtr batch1, IntPtr batch2, IntPtr mat, float beta,
@@ -3097,85 +2035,6 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_bitwise_and(IntPtr tensor, IntPtr other);
-
-        public TorchTensor bitwise_and(TorchTensor other)
-        {
-            var res = THSTensor_bitwise_and(handle, other.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_bitwise_and_(IntPtr tensor, IntPtr other);
-
-        public TorchTensor bitwise_and_(TorchTensor other)
-        {
-            var res = THSTensor_bitwise_and_(handle, other.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_bitwise_not(IntPtr tensor);
-
-        public TorchTensor bitwise_not()
-        {
-            var res = THSTensor_bitwise_not(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_bitwise_not_(IntPtr tensor);
-
-        public TorchTensor bitwise_not_(TorchTensor other)
-        {
-            var res = THSTensor_bitwise_not_(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_bitwise_or(IntPtr tensor, IntPtr other);
-
-        public TorchTensor bitwise_or(TorchTensor other)
-        {
-            var res = THSTensor_bitwise_or(handle, other.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_bitwise_or_(IntPtr tensor, IntPtr other);
-
-        public TorchTensor bitwise_or_(TorchTensor other)
-        {
-            var res = THSTensor_bitwise_or_(handle, other.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_bitwise_xor(IntPtr tensor, IntPtr other);
-
-        public TorchTensor bitwise_xor(TorchTensor other)
-        {
-            var res = THSTensor_bitwise_xor(handle, other.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_bitwise_xor_(IntPtr tensor, IntPtr other);
-
-        public TorchTensor bitwise_xor_(TorchTensor other)
-        {
-            var res = THSTensor_bitwise_xor_(handle, other.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
 
         public TorchTensor @bool() => this.to_type(ScalarType.Bool);
 
@@ -3190,116 +2049,6 @@ namespace TorchSharp.Tensor
         public TorchTensor @float() => this.to_type(ScalarType.Float32);
 
         public TorchTensor @double() => this.to_type(ScalarType.Float64);
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_logical_and(IntPtr tensor, IntPtr other);
-
-        public TorchTensor logical_and(TorchTensor other)
-        {
-            var res = THSTensor_logical_and(handle, other.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_logical_and_(IntPtr tensor, IntPtr other);
-
-        public TorchTensor logical_and_(TorchTensor other)
-        {
-            var res = THSTensor_logical_and_(handle, other.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_logical_not(IntPtr tensor);
-
-        public TorchTensor logical_not()
-        {
-            var res = THSTensor_logical_not(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_logical_not_(IntPtr tensor);
-
-        public TorchTensor logical_not_(TorchTensor other)
-        {
-            var res = THSTensor_logical_not_(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_logical_or(IntPtr tensor, IntPtr other);
-
-        public TorchTensor logical_or(TorchTensor other)
-        {
-            var res = THSTensor_logical_or(handle, other.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_logical_or_(IntPtr tensor, IntPtr other);
-
-        public TorchTensor logical_or_(TorchTensor other)
-        {
-            var res = THSTensor_logical_or_(handle, other.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_logical_xor(IntPtr tensor, IntPtr other);
-
-        public TorchTensor logical_xor(TorchTensor other)
-        {
-            var res = THSTensor_logical_xor(handle, other.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_logical_xor_(IntPtr tensor, IntPtr other);
-
-        public TorchTensor logical_xor_(TorchTensor other)
-        {
-            var res = THSTensor_logical_xor_(handle, other.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_cholesky(IntPtr input, bool upper);
-
-        public TorchTensor cholesky(bool upper = false)
-        {
-            var res = THSTensor_cholesky(handle, upper);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_cholesky_inverse(IntPtr input, bool upper);
-
-        public TorchTensor cholesky_inverse(bool upper = false)
-        {
-            var res = THSTensor_cholesky_inverse(handle, upper);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_cholesky_solve(IntPtr input, IntPtr input2, bool upper);
-
-        public TorchTensor cholesky_solve(TorchTensor input2, bool upper = false)
-        {
-            var res = THSTensor_cholesky_solve(handle, input2.Handle, upper);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_clamp(IntPtr input, IntPtr min, IntPtr max);
@@ -3359,20 +2108,6 @@ namespace TorchSharp.Tensor
         public TorchTensor clamp_min_(TorchScalar min)
         {
             var res = THSTensor_clamp_min_(handle, min.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_cross(IntPtr input, IntPtr other, long dim);
-
-        /// <summary>
-        /// Returns the cross product of vectors in dimension dim of input and other.
-        /// input and other must have the same size, and the size of their dim dimension should be 3.
-        /// </summary>
-        public TorchTensor cross(TorchScalar other, long dim)
-        {
-            var res = THSTensor_cross(handle, other.Handle, dim);
             if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
             return new TorchTensor(res);
         }
@@ -3466,49 +2201,6 @@ namespace TorchSharp.Tensor
         public TorchTensor diagonal(long offset = 0, long dim1 = 0, long dim2 = 0)
         {
             var res = THSTensor_diagonal(handle, offset, dim1, dim2);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_div(IntPtr tensor, IntPtr trg);
-
-        public TorchTensor div(TorchTensor target)
-        {
-            var res = THSTensor_div(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        public TorchTensor divide(TorchTensor target) => div(target);
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_div_scalar(IntPtr tensor, IntPtr trg);
-
-        public TorchTensor div(TorchScalar target)
-        {
-            var res = THSTensor_div_scalar(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-        public TorchTensor divide(TorchScalar target) => div(target);
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_div_(IntPtr tensor, IntPtr trg);
-
-        public TorchTensor div_(TorchTensor target)
-        {
-            var res = THSTensor_div_(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_div_scalar_(IntPtr tensor, IntPtr trg);
-
-        public TorchTensor div_(TorchScalar target)
-        {
-            var res = THSTensor_div_scalar_(handle, target.Handle);
             if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
             return new TorchTensor(res);
         }
@@ -3634,136 +2326,6 @@ namespace TorchSharp.Tensor
             var res = THSTensor_allclose(handle, target.Handle, rtol, atol, equal_nan);
             Torch.CheckForErrors();
             return res;
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_exp(IntPtr tensor);
-
-        public TorchTensor exp()
-        {
-            var res = THSTensor_exp(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_exp_(IntPtr tensor);
-
-        public TorchTensor exp_()
-        {
-            var res = THSTensor_exp_(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_exp2(IntPtr tensor);
-
-        public TorchTensor exp2()
-        {
-            var res = THSTensor_exp2(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_expm1(IntPtr tensor);
-
-        public TorchTensor expm1()
-        {
-            var res = THSTensor_expm1(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_expm1_(IntPtr tensor);
-
-        public TorchTensor expm1_()
-        {
-            var res = THSTensor_expm1_(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_fft(IntPtr tensor, long n, long dim, [MarshalAs(UnmanagedType.LPStr)] string norm);
-
-        public TorchTensor fft(long? n, long dim = -1, string norm = "backward")
-        {
-            var res = THSTensor_fft(handle, n.GetValueOrDefault(-1), dim, norm);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_ifft(IntPtr tensor, long n, long dim, [MarshalAs(UnmanagedType.LPStr)] string norm);
-
-        public TorchTensor ifft(long? n, long dim = -1, string norm = "backward")
-        {
-            var res = THSTensor_ifft(handle, n.GetValueOrDefault(-1), dim, norm);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_irfft(IntPtr tensor, long n, long dim, [MarshalAs(UnmanagedType.LPStr)] string norm);
-
-        public TorchTensor irfft(long? n, long dim = -1, string norm = "backward")
-        {
-            var res = THSTensor_irfft(handle, n.GetValueOrDefault(-1), dim, norm);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_rfft(IntPtr tensor, long n, long dim, [MarshalAs(UnmanagedType.LPStr)] string norm);
-
-        public TorchTensor rfft(long? n, long dim = -1, string norm = "backward")
-        {
-            var res = THSTensor_rfft(handle, n.GetValueOrDefault(-1), dim, norm);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_frac(IntPtr tensor);
-
-        public TorchTensor frac()
-        {
-            var res = THSTensor_frac(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_frac_(IntPtr tensor);
-
-        public TorchTensor frac_()
-        {
-            var res = THSTensor_frac_(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_gcd(IntPtr tensor, IntPtr other);
-
-        public TorchTensor gcd(TorchTensor other)
-        {
-            var res = THSTensor_gcd(handle, other.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_gcd_(IntPtr tensor, IntPtr other);
-
-        public TorchTensor gcd_(TorchTensor other)
-        {
-            var res = THSTensor_gcd_(handle, other.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
         }
 
         [DllImport("LibTorchSharp")]
@@ -3933,26 +2495,6 @@ namespace TorchSharp.Tensor
         }
 
         [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_log(IntPtr tensor);
-
-        public TorchTensor log()
-        {
-            var res = THSTensor_log(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_log_(IntPtr tensor);
-
-        public TorchTensor log_()
-        {
-            var res = THSTensor_log_(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_lt(IntPtr tensor, IntPtr trg);
 
         public TorchTensor lt(TorchTensor target)
@@ -4036,16 +2578,6 @@ namespace TorchSharp.Tensor
         }
 
         [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_matmul(IntPtr tensor, IntPtr target);
-
-        public TorchTensor matmul(TorchTensor target)
-        {
-            var res = THSTensor_matmul(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
         static extern void THSTensor_topk(IntPtr tensor, AllocatePinnedArray allocator, int k,
             long dimension, bool largest, bool sorted);
 
@@ -4062,30 +2594,6 @@ namespace TorchSharp.Tensor
 
             return (new TorchTensor(ptrArray[0]), new TorchTensor(ptrArray[1]));
         }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_trunc(IntPtr tensor);
-
-        public TorchTensor trunc()
-        {
-            var res = THSTensor_trunc(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        public TorchTensor fix() => trunc();
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_trunc_(IntPtr tensor);
-
-        public TorchTensor trunc_()
-        {
-            var res = THSTensor_trunc_(handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        public TorchTensor fix_() => trunc_();
 
 
         [DllImport("LibTorchSharp")]
@@ -4413,26 +2921,6 @@ namespace TorchSharp.Tensor
         }
 
         [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_mm(IntPtr tensor, IntPtr target);
-
-        public TorchTensor mm(TorchTensor target)
-        {
-            var res = THSTensor_mm(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_mv(IntPtr tensor, IntPtr target);
-
-        public TorchTensor mv(TorchTensor target)
-        {
-            var res = THSTensor_mv(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_msort(IntPtr tensor);
 
         public TorchTensor msort()
@@ -4442,60 +2930,6 @@ namespace TorchSharp.Tensor
             return new TorchTensor(res);
         }
 
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_vdot(IntPtr tensor, IntPtr target);
-
-        public TorchTensor vdot(TorchTensor target)
-        {
-            if (shape.Length != 1 || target.shape.Length != 1 || shape[0] != target.shape[0]) throw new InvalidOperationException("vdot arguments must have the same shape.");
-            var res = THSTensor_vdot(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_mul(IntPtr tensor, IntPtr target);
-
-        public TorchTensor mul(TorchTensor target)
-        {
-            var res = THSTensor_mul(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        public TorchTensor multiply(TorchTensor target) => mul(target);
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_mul_scalar(IntPtr tensor, IntPtr scalar);
-
-        public TorchTensor mul(TorchScalar scalar)
-        {
-            var res = THSTensor_mul_scalar(handle, scalar.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        public TorchTensor multiply(TorchScalar target) => mul(target);
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_mul_(IntPtr tensor, IntPtr target);
-
-        public TorchTensor mul_(TorchTensor target)
-        {
-            var res = THSTensor_mul_(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_mul_scalar_(IntPtr tensor, IntPtr target);
-
-        public TorchTensor mul_(TorchScalar target)
-        {
-            var res = THSTensor_mul_scalar_(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
 
         [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_ne(IntPtr tensor, IntPtr trg);
@@ -4602,102 +3036,11 @@ namespace TorchSharp.Tensor
         }
 
         [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_pow(IntPtr tensor, IntPtr exponent);
-
-        public TorchTensor pow(TorchTensor exponent)
-        {
-            var res = THSTensor_pow(handle, exponent.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_pow_(IntPtr tensor, IntPtr exponent);
-
-        public TorchTensor pow_(TorchTensor exponent)
-        {
-            var res = THSTensor_pow_(handle, exponent.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_pow_scalar(IntPtr tensor, IntPtr scalar);
-
-        public TorchTensor pow(TorchScalar scalar)
-        {
-            var res = THSTensor_pow_scalar(handle, scalar.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_pow_scalar_(IntPtr tensor, IntPtr scalar);
-
-        public TorchTensor pow_(TorchScalar scalar)
-        {
-            var res = THSTensor_pow_scalar_(handle, scalar.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
         static extern IntPtr THSTensor_prelu(IntPtr tensor, IntPtr trg);
 
         public TorchTensor prelu(TorchTensor target)
         {
             var res = THSTensor_prelu(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_remainder(IntPtr tensor, IntPtr trg);
-
-        public TorchTensor remainder(TorchTensor target)
-        {
-            var res = THSTensor_remainder(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_remainder_(IntPtr tensor, IntPtr trg);
-
-        public TorchTensor remainder_(TorchTensor target)
-        {
-            var res = THSTensor_remainder_(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_remainder_scalar(IntPtr tensor, IntPtr scalar);
-
-        public TorchTensor remainder(TorchScalar scalar)
-        {
-            var res = THSTensor_remainder_scalar(handle, scalar.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_remainder_scalar_(IntPtr tensor, IntPtr scalar);
-
-        public TorchTensor remainder_(TorchScalar scalar)
-        {
-            var res = THSTensor_remainder_scalar_(handle, scalar.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_float_power(IntPtr tensor, IntPtr trg);
-
-        public TorchTensor float_power(TorchTensor target)
-        {
-            var res = THSTensor_float_power(handle, target.Handle);
             if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
             return new TorchTensor(res);
         }
@@ -4718,46 +3061,6 @@ namespace TorchSharp.Tensor
         public TorchTensor fmin(TorchTensor target)
         {
             var res = THSTensor_fmin(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_fmod(IntPtr tensor, IntPtr trg);
-
-        public TorchTensor fmod(TorchTensor target)
-        {
-            var res = THSTensor_fmod(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_fmod_(IntPtr tensor, IntPtr trg);
-
-        public TorchTensor fmod_(TorchTensor target)
-        {
-            var res = THSTensor_fmod_(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_fmod_scalar(IntPtr tensor, IntPtr scalar);
-
-        public TorchTensor fmod(TorchScalar scalar)
-        {
-            var res = THSTensor_fmod_scalar(handle, scalar.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_fmod_scalar_(IntPtr tensor, IntPtr scalar);
-
-        public TorchTensor fmod_(TorchScalar scalar)
-        {
-            var res = THSTensor_fmod_scalar_(handle, scalar.Handle);
             if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
             return new TorchTensor(res);
         }
@@ -4814,46 +3117,6 @@ namespace TorchSharp.Tensor
                     return new TorchTensor(res);
                 }
             }
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_sub(IntPtr tensor, IntPtr trg);
-
-        public TorchTensor sub(TorchTensor target)
-        {
-            var res = THSTensor_sub(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_sub_scalar(IntPtr tensor, IntPtr trg);
-
-        public TorchTensor sub(TorchScalar target)
-        {
-            var res = THSTensor_sub_scalar(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_sub_(IntPtr tensor, IntPtr trg);
-
-        public TorchTensor sub_(TorchTensor target)
-        {
-            var res = THSTensor_sub_(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_sub_scalar_(IntPtr tensor, IntPtr trg);
-
-        public TorchTensor sub_(TorchScalar target)
-        {
-            var res = THSTensor_sub_scalar_(handle, target.Handle);
-            if (res == IntPtr.Zero) { Torch.CheckForErrors(); }
-            return new TorchTensor(res);
         }
 
         [DllImport("LibTorchSharp")]
@@ -5241,22 +3504,6 @@ namespace TorchSharp.Tensor
                     return new TorchTensor(res);
                 }
             }
-        }
-
-        [DllImport("LibTorchSharp")]
-        static extern IntPtr THSTensor_xlogy(IntPtr tensor, IntPtr trg);
-
-        /// <summary>
-        /// Computes x * log(y)
-        /// </summary>
-        /// <param name="target"></param>
-        /// <returns></returns>
-        public TorchTensor xlogy(TorchTensor target)
-        {
-            var res = THSTensor_xlogy(handle, target.Handle);
-            if (res == IntPtr.Zero)
-                Torch.CheckForErrors();
-            return new TorchTensor(res);
         }
 
         [DllImport("LibTorchSharp")]
@@ -6441,66 +4688,6 @@ namespace TorchSharp.Tensor
         public static TorchTensor operator !=(TorchScalar left, TorchTensor right)
         {
             return right.ne(left);
-        }
-
-        public static TorchTensor operator +(TorchTensor left, TorchTensor right)
-        {
-            return left.add(right);
-        }
-
-        public static TorchTensor operator +(TorchTensor left, TorchScalar right)
-        {
-            return left.add(right);
-        }
-
-        public static TorchTensor operator +(TorchScalar left, TorchTensor right)
-        {
-            return right.add(left);
-        }
-
-        public static TorchTensor operator *(TorchTensor left, TorchTensor right)
-        {
-            return left.mul(right);
-        }
-
-        public static TorchTensor operator *(TorchTensor left, TorchScalar right)
-        {
-            return left.mul(right);
-        }
-
-        public static TorchTensor operator *(TorchScalar left, TorchTensor right)
-        {
-            return right.mul(left);
-        }
-
-        public static TorchTensor operator -(TorchTensor left, TorchTensor right)
-        {
-            return left.sub(right);
-        }
-
-        public static TorchTensor operator -(TorchTensor left, TorchScalar right)
-        {
-            return left.sub(right);
-        }
-
-        public static TorchTensor operator /(TorchTensor left, TorchTensor right)
-        {
-            return left.div(right);
-        }
-
-        public static TorchTensor operator /(TorchTensor left, TorchScalar right)
-        {
-            return left.div(right);
-        }
-
-        public static TorchTensor operator %(TorchTensor left, TorchTensor right)
-        {
-            return left.remainder(right);
-        }
-
-        public static TorchTensor operator %(TorchTensor left, TorchScalar right)
-        {
-            return left.remainder(right);
         }
 
         public static TorchTensor operator <(TorchTensor left, TorchTensor right)
